BGK1D1V訓練スクリプトの損失関数loss_scaled_dnu呼び出しから入力テンソル(pred/y/x)の明示的float32キャストを削除し、関数内部のキャスト処理に一本化

+++ src/kineticEQ/CNN/BGK1D1V/train.py
@@ -185,7 +185,7 @@ def main():
             with torch.amp.autocast("cuda", enabled=use_amp):
                 pred = model(x)
                 loss = loss_scaled_dnu(
-                    pred.float(), y.float(), x.float(),
+                    pred, y, x,
                     eps=1e-6,
                     u_eps=float(args.u_eps),
                     s_min=float(args.s_min),
